requirements.txt
SQLAlchemy~=2.0.36
Flask~=3.1.0
pandas~=2.2.3
Werkzeug~=3.1.3
python-dotenv~=1.0.1
Flask-Bcrypt~=1.0.1
Flask-RESTful~=0.3.9